refactor(rule-criteria): hoist ConditionParamsFields to a shared feature module
Moves `ConditionParamsFields`, `MultiSelectField`, and
`getDefaultConditionParams` out of `features/library-cleanup/components/`
into a new `features/rule-criteria/components/` module. Both
library-cleanup and auto-tag consume from the new location via
relative imports across the feature-folder boundary — no longer
reaching into another feature's `components/` directory.

This addresses code-review finding #6 from PR #396 (auto-tagger
webhook): the cross-feature import made `condition-params-fields.tsx`
a de-facto public API of `library-cleanup` without it being labeled
as such.

Pure refactor — no behavior change.

What moved
----------
- `features/library-cleanup/components/condition-params-fields.tsx`
  → `features/rule-criteria/components/condition-params-fields.tsx`
- `features/library-cleanup/components/multi-select-field.tsx`
  → `features/rule-criteria/components/multi-select-field.tsx`

What's updated
--------------
- `features/library-cleanup/components/cleanup-rule-dialog.tsx` —
  imports updated to `../../rule-criteria/components/...`
- `features/auto-tag/components/rule-dialog.tsx` — same
- `features/library-cleanup/lib/__tests__/condition-params.test.ts` —
  test import path updated

Validation
----------
- `pnpm --filter @arr/web exec tsc --noEmit` — clean
- `pnpm --filter @arr/api run test` — 1408 passing (no regression)
- `pnpm run lint` — no new findings
- Browser-verified: library-cleanup page renders, "Add Rule" dialog
  opens with ConditionParamsFields present and rule-type-specific
  params rendering correctly. 0 console errors.

Refs: deferred follow-up #6 from PR #396 review
